fix(cms): Refactored Pricing and Studio to use CMS content
